Finishing change architecture
