+++ components/polymerBases/routersPolymerBases.py
@@ -1,6 +1,6 @@
-from fastapi import APIRouter, Cookie, Response, Depends, HTTPException
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
-from db import crud_old, schemas, models
+from components.polymerBases import models
 from db.database import get_db
 from configparser import ConfigParser
