+++ components/vacuumSystem/routersVacuumSystem.py
@@ -1,8 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from starlette.responses import JSONResponse
-
-from db import crud_old, schemas
+from components.vacuumSystem import schemas, crud
 from db.database import get_db
 from configparser import ConfigParser
